+++ src/Yasumi/data/translations/secondNewYearsDay.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * This file is part of the Yasumi package.
+ *
+ * Copyright (c) 2015 - 2019 AzuyaLabs
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Sacha Telgenhof <me@sachatelgenhof.com>
+ */
+
+// Translations for New Year's Day
+return [
+    'da_DK' => '2. nytårsdag',
+    'en_GB' => '2nd January',
+];