+++ tests/UnitedKingdom/Wales/NewYearsDayTest.php
@@ -1,0 +1,126 @@
+<?php
+/**
+ * This file is part of the Yasumi package.
+ *
+ * Copyright (c) 2015 - 2019 AzuyaLabs
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Sacha Telgenhof <me@sachatelgenhof.com>
+ */
+
+namespace Yasumi\tests\UnitedKingdom\Wales;
+
+use DateTime;
+use DateTimeZone;
+use Yasumi\Holiday;
+use Yasumi\tests\YasumiTestCaseInterface;
+
+/**
+ * Class for testing New Years Day in Wales.
+ */
+class NewYearsDayTest extends WalesBaseTestCase implements YasumiTestCaseInterface
+{
+    /**
+     * The year in which the holiday was first established
+     */
+    public const ESTABLISHMENT_YEAR = 1871;
+
+    /**
+     * The year in which the holiday was adjusted
+     */
+    public const ADJUSTMENT_YEAR = 1974;
+
+    /**
+     * The name of the holiday to be tested
+     */
+    public const HOLIDAY = 'newYearsDay';
+
+    /**
+     * Tests the holiday defined in this test on or after establishment.
+     *
+     * @dataProvider HolidayDataProvider
+     *
+     * @param int    $year     the year for which the holiday defined in this test needs to be tested
+     * @param string $expected the expected date
+     *
+     * @throws \ReflectionException
+     * @throws \Exception
+     */
+    public function testHolidayOnAfterEstablishment($year, $expected)
+    {
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            $year,
+            new DateTime($expected, new DateTimeZone(self::TIMEZONE))
+        );
+    }
+
+    /**
+     * Tests the holiday defined in this test before establishment.
+     * @throws \ReflectionException
+     */
+    public function testHolidayBeforeEstablishment()
+    {
+        $this->assertNotHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(1000, self::ESTABLISHMENT_YEAR - 1)
+        );
+    }
+
+    /**
+     * Tests that the holiday defined in this test is of the type 'observance' before the year it was changed.
+     * @throws \ReflectionException
+     */
+    public function testHolidayIsObservedTypeBeforeChange()
+    {
+        $this->assertHolidayType(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(self::ESTABLISHMENT_YEAR, self::ADJUSTMENT_YEAR - 1),
+            Holiday::TYPE_OBSERVANCE
+        );
+    }
+
+    /**
+     * Returns a list of random test dates used for assertion of the holiday defined in this test
+     *
+     * @return array list of test dates for the holiday defined in this test
+     * @throws \Exception
+     */
+    public function HolidayDataProvider(): array
+    {
+        return $this->generateRandomDatesWithHolidayMovedToMonday(01, 01, self::TIMEZONE, 10, self::ESTABLISHMENT_YEAR);
+    }
+
+    /**
+     * Tests the translated name of the holiday defined in this test.
+     * @throws \ReflectionException
+     */
+    public function testTranslation(): void
+    {
+        $this->assertTranslatedHolidayName(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(self::ESTABLISHMENT_YEAR),
+            [self::LOCALE => 'New Year\'s Day']
+        );
+    }
+
+    /**
+     * Tests type of the holiday defined in this test.
+     * @throws \ReflectionException
+     */
+    public function testHolidayType(): void
+    {
+        $this->assertHolidayType(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(self::ADJUSTMENT_YEAR + 1),
+            Holiday::TYPE_BANK
+        );
+    }
+}